cadillac

+++ Streetile/Fusca.h
@@ -23,7 +23,7 @@ public:
 	void DefaultSpeed() {
 		speed = 2;
 	}
-	void setSpeed(int speed) {
+	void setSpeed(float speed) {
 		this->speed = speed;
 	}
