+++ Streetile/Perua.h
@@ -21,7 +21,7 @@ public:
 	void DefaultSpeed() {
 		speed = 2;
 	}
-	void setSpeed(int speed) {
+	void setSpeed(float speed) {
 		this->speed = speed;
 	}
 };
